requirements.txt
Flask==0.10.1
Flask-OAuthlib==0.7.0
Flask-SQLAlchemy==2.0
Flask-WTF==0.10.2
Jinja2==2.7.3
MarkupSafe==0.23
MySQL-python==1.2.5
SQLAlchemy==0.9.7
WTForms==2.0.1
Werkzeug==0.9.6
WTAlchemy==0.1b3
argparse==1.2.1
itsdangerous==0.24
## THIS IS CRUCIAL! need oauthlib = 0.6.0; otherwise shit breaks...
oauthlib==0.6.0
wsgiref==0.1.2
